PromptEngineering/LLM_tests_framework/src/frameworks/custom_metrics.py
from dataclasses import dataclass
from typing import Dict, List, Optional, Any
import pandas as pd
import numpy as np
import json
from concurrent.futures import ThreadPoolExecutor

@dataclass
class ModelMetrics:
    name: str
    performance_scores: Dict[str, float]
    cost_per_1k_tokens: float
    avg_latency_ms: float
    max_context_length: int
    supported_features: List[str]
    license_type: str
    hosting_options: List[str]

class CustomMetrics:
    def __init__(self, config_path: str):
        self.config = self._load_config(config_path)
        self.models: Dict[str, ModelMetrics] = {}
        self.benchmark_datasets = self._load_benchmark_datasets()
        self.evaluation_criteria = self._setup_evaluation_criteria()

    def _load_config(self, config_path: str) -> Dict:
        """Load configuration from JSON file"""
        with open(config_path, 'r') as file:
            return json.load(file)

    def _load_benchmark_datasets(self):
        return {
            'mmlu': self._load_mmlu(),
            'hellaswag': self._load_hellaswag(),
            'truthfulqa': self._load_truthfulqa(),
            'humaneval': self._load_humaneval(),
            'custom': self._load_custom_benchmarks()
        }

    def evaluate_model(self, model_name: str, model_config: Dict) -> ModelMetrics:
        """Comprehensive model evaluation"""
        with ThreadPoolExecutor() as executor:
            # Parallel tests on different datasets
            performance_futures = {
                dataset: executor.submit(self._evaluate_on_dataset, 
                                      model_name, 
                                      dataset_data)
                for dataset, dataset_data in self.benchmark_datasets.items()
            }

            # Performance tests
            latency_future = executor.submit(self._measure_latency, model_name)
            cost_future = executor.submit(self._estimate_costs, model_name)

            # Collecting results
            performance_scores = {
                dataset: future.result()
                for dataset, future in performance_futures.items()
            }

            return ModelMetrics(
                name=model_name,
                performance_scores=performance_scores,
                cost_per_1k_tokens=cost_future.result(),
                avg_latency_ms=latency_future.result(),
                max_context_length=model_config['max_context_length'],
                supported_features=model_config['features'],
                license_type=model_config['license'],
                hosting_options=model_config['hosting']
            )

    def calculate_model_score(self, metrics: ModelMetrics) -> float:
        """Calculate final model score"""
        weights = self.evaluation_criteria['weights']

        performance_score = np.mean([
            score * weights['performance']
            for score in metrics.performance_scores.values()
        ])

        cost_score = (1.0 / metrics.cost_per_1k_tokens) * weights['cost']
        latency_score = (1.0 / metrics.avg_latency_ms) * weights['latency']

        feature_score = len(metrics.supported_features) * weights['features']

        return performance_score + cost_score + latency_score + feature_score

    def generate_recommendation(self) -> Dict:
        """Generate model selection recommendation"""
        model_scores = {
            name: self.calculate_model_score(metrics)
            for name, metrics in self.models.items()
        }

        best_model = max(model_scores.items(), key=lambda x: x[1])[0]

        return {
            'recommended_model': best_model,
            'detailed_scores': model_scores,
            'analysis': self._generate_analysis(best_model),
            'trade_offs': self._analyze_trade_offs(best_model)
        }

    def _analyze_trade_offs(self, model_name: str) -> Dict:
        """Analyze trade-offs for the selected model"""
        metrics = self.models[model_name]

        return {
            'performance_vs_cost': {
                'performance': metrics.performance_scores,
                'cost_analysis': self._analyze_cost_efficiency(metrics),
                'recommendations': self._generate_cost_recommendations(metrics)
            },
            'speed_vs_accuracy': {
                'latency': metrics.avg_latency_ms,
                'accuracy': np.mean(list(metrics.performance_scores.values())),
                'recommendations': self._generate_optimization_recommendations(metrics)
            }
        }

    def _load_mmlu(self):
        # Load MMLU dataset
        pass

    def _load_hellaswag(self):
        # Load HellaSwag dataset
        pass

    def _load_truthfulqa(self):
        # Load TruthfulQA dataset
        pass

    def _load_humaneval(self):
        # Load HumanEval dataset
        pass

    def _load_custom_benchmarks(self):
        # Load custom benchmarks
        pass

    def _evaluate_on_dataset(self, model_name: str, dataset_data: Any) -> float:
        # Evaluate model on a specific dataset
        pass

    def _measure_latency(self, model_name: str) -> float:
        # Measure model latency
        pass

    def _estimate_costs(self, model_name: str) -> float:
        # Estimate model costs
        pass

    def _setup_evaluation_criteria(self) -> Dict:
        # Setup evaluation criteria
        pass

    def _generate_analysis(self, model_name: str) -> Dict:
        # Generate analysis for the model
        pass

    def _analyze_cost_efficiency(self, metrics: ModelMetrics) -> Dict:
        # Analyze cost efficiency
        pass

    def _generate_cost_recommendations(self, metrics: ModelMetrics) -> List[str]:
        # Generate cost recommendations
        pass

    def _generate_optimization_recommendations(self, metrics: ModelMetrics) -> List[str]:
        # Generate optimization recommendations
        pass